// src/app/features/mastertables/tblUserDetail/models/tblUserDetail.model.ts
export interface TblUserDetail {
	fldId: number;
	fldFKUser: number;
	fldFKMenuMaster: number;
	fldCondition: string;
	fldFKMenuCategory: number;
	fldFKRole: number;
	fldRoleName: string;
	fldRole: string;
	fldMenuRoleCode: string;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
	fldModifiedBy: number;
	fldModifiedDt: Date;
	fldDeletedBy: number;
	fldDeletedDt: Date;
}