// src/app/features/mastertables/tblUserDetail/models/tblUserDetail-Update.model.ts
export interface TblUserDetailUpdate {
	fldId: number;
	fldFKUser: number;
	fldFKMenuMaster: number;
	fldCondition: string;
	fldFKMenuCategory: number;
	fldFKRole: number;
	fldRoleName: string;
	fldRole: string;
	fldMenuRoleCode: string;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
	fldModifiedBy: number;
	fldModifiedDt: Date;
}